requirements.txt
garminconnect>=0.2.19
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
pytest>=7.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyyaml>=6.0
portalocker>=2.8.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
alembic>=1.13.0
fastapi-users[sqlalchemy]>=13.0.0
cryptography>=42.0.0
azure-identity>=1.15.0
azure-keyvault-keys>=4.9.0
gunicorn>=22.0.0
PyJWT>=2.8.0
httpx>=0.27.0
